Działa dodawanie kategorii, usuwanie, pobieranie z bazy oraz synchronizacja z mailService
widok sie binduje ok

+++ StudentMailOrganizer/StudentMailOrganizer/DAL/DatabaseManager.cs
@@ -1,5 +1,6 @@
 ﻿using StudentMailOrganizer.Infrastructure;
 using StudentMailOrganizer.Models;
+using StudentMailOrganizer.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -19,38 +20,65 @@ namespace StudentMailOrganizer.DAL
             db = new MailContext();
         }
 
-        public IEnumerable<Category> Synchronize()
+        public SynchMailViewModel Synchronize()
         {
             var emails = _mailer.GetAllMessages().ToList();
-            db.Database.ExecuteSqlCommand("TRUNCATE TABLE dbo.MailMessages");
+            db.Database.ExecuteSqlCommand("TRUNCATE TABLE dbo.MailMessageCategories;");
+            db.Database.ExecuteSqlCommand("DELETE FROM dbo.MailMessages WHERE 1=1;");
             foreach (var email in emails)
             {
                 db.Emails.Add(email);
             }
+            var categories = SortMessages(emails);
             db.SaveChanges();
 
-            var sortedMails = SortMessages(emails);
+            SynchMailViewModel vm = new SynchMailViewModel()
+            {
+                Emails = emails,
+                Categories = categories.ToList()
+            };
 
-            return sortedMails;
+            return vm;
+        }
+
+        public SynchMailViewModel GetDataFromDatabase()
+        {
+            var vm = new SynchMailViewModel();
+            vm.Categories = GetCategories().ToList();
+            vm.Emails = GetMessages().ToList();
+            return vm;
         }
 
         public IEnumerable<Category> GetCategories()
         {
             return db.Categories;
         }
+        public IEnumerable<MailMessage> GetMessages()
+        {
+            return db.Emails;
+        }
 
         public bool AddCategory(Category category)
         {
-            try
+            //try
+            //{
+            category.Mails = new List<MailMessage>();
+            foreach (var filter in category.AcceptedEmails)
             {
-                db.Categories.Add(category);
-                db.SaveChanges();
-                return true;
+                var emails = db.Emails.Where(x => x.Sender == filter.Email);
+                foreach (var email in emails)
+                {
+                    category.Mails.Add(email);
+                }
             }
-            catch (Exception)
-            {
-                return false;
-            }
+            db.Categories.Add(category);
+            db.SaveChanges();
+            return true;
+            //}
+            //catch (Exception)
+            //{
+            //    return false;
+            //}
         }
 
         public bool EditCategory(Category category)
@@ -58,17 +86,17 @@ namespace StudentMailOrganizer.DAL
             var cat = db.Categories.FirstOrDefault(x => x.CategoryId == category.CategoryId);
             if (cat != null)
             {
-                try
-                {
-                    cat.Name = category.Name;
-                    cat.AcceptedEmails = category.AcceptedEmails;
-                    db.SaveChanges();
-                    return true;
-                }
-                catch (Exception)
-                {
-                    return false;
-                }
+                //try
+                //{
+                cat.Name = category.Name;
+                cat.AcceptedEmails = category.AcceptedEmails;
+                db.SaveChanges();
+                return true;
+                //}
+                //catch (Exception)
+                //{
+                //    return false;
+                //}
             }
             else
             {
@@ -78,53 +106,34 @@ namespace StudentMailOrganizer.DAL
 
         public bool RemoveCategory(Category category)
         {
-            try
-            {
-                db.Categories.Remove(category);
-                db.SaveChanges();
-                return true;
-            }
-            catch (Exception)
-            {
-                return false;
-            }
+            //try
+            //{
+            if (category.CategoryId == -1) return true;
+            category.Mails.Clear();
+            db.Categories.Remove(category);
+            db.SaveChanges();
+            return true;
+            //}
+            //catch (Exception)
+            //{
+            //    return false;
+            //}
         }
 
-        public IEnumerable<MailMessage> GetAllMessages()
-        {
-            return db.Emails;
-        }
 
         private IEnumerable<Category> SortMessages(IEnumerable<MailMessage> emails)
         {
             var categories = GetCategories().ToList();
             foreach (var cat in categories)
             {
+                var filter = cat.AcceptedEmails.Select(x => x.Email).ToList();
                 foreach (var email in emails)
                 {
-                    var filter = cat.AcceptedEmails.Select(x => x.Email);
                     if (filter.Contains(email.Sender))
                         cat.Mails.Add(email);
                 }
             }
             return categories;
         }
-        public  IEnumerable<Category> GetSortedMessages()
-        {
-            var emails = GetAllMessages().ToList();
-            var categories = GetCategories().ToList();
-            foreach (var cat in categories)
-            {
-                foreach (var email in emails)
-                {
-                    var filter = cat.AcceptedEmails.Select(x => x.Email);
-                    if (filter.Contains(email.Sender))
-                        cat.Mails.Add(email);
-                }
-            }
-            return categories;
-        }
-
-
     }
 }
